refactor: merge ProviderRepositoryFirestoneTest into ProviderRepositoryFirestoreTest
- Fixed typo in test file name (Firestone -> Firestore)
- Merged returnProvider test cases from old file
- Adapted tests to use setupBasicProviderFields helper
- Removed duplicate test file
